Preliminary Spinner has started working
Basic Spinner is installed on home page. Can change the activity to all the other pages including the archives website. I will add more of the spinners functionality soon. Also, there is some formatting changes that fix the layout so that it changes correctly when switching between portrait and landscape.